Add convert_to_pddl, easier cache skipping, input pdkbdll string.
convert_to_pddl makes it easier to just get the PDDL files from a PDKBDLL file when you don't actually want to do the planning.
A couple of functions which required a pdkbdll file path as input will now accept a pdkbdll string too; this is sometimes useful for interactive purposes so you don't have to write then read a file.